Change the package name of tomcat it module
