+++ it/spring/boot2-tomcat9/src/test/java/com/linecorp/armeria/spring/tomcat/SpringTomcatApplicationItTest.java
@@ -14,7 +14,7 @@
  * under the License.
  */
 
-package com.linecorp.armeria.spring.tomcat.demo;
+package com.linecorp.armeria.spring.tomcat;
 
 import static org.assertj.core.api.Assertions.assertThat;
